- SEO changes for heading and sections tags and structure - some minor code refactoring - added some new meta tags for better SEO - added dynamic meta tags - https://schema.org js object in header for SEO purposes
