+++ public/index.html
@@ -3,12 +3,35 @@
   <head>
     <meta charset="utf-8">
     <meta http-equiv="X-UA-Compatible" content="IE=edge">
-    <meta name="viewport" content="height=device-height, width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no, target-densitydpi=device-dpi">
+    <meta name="viewport" content="height=device-height, width=device-width, initial-scale=1.0, minimum-scale=1.0, maximum-scale=1.0, user-scalable=no, target-densitydpi=device-dpi" />
+    <meta property="og:locale" content="en_UK" />
+    <meta property="og:type" content="website" />
+    <meta property="og:url" content="https://risidio.com/" />
+    <meta property="og:site_name" content="Risidio" />
     <link rel="icon" href="<%= BASE_URL %>img/icons/favicon_dice16x161.png">
-    <title>Risidio</title>
     <script src="https://kit.fontawesome.com/5885eabe05.js" crossorigin="anonymous"></script>
     <script type="module" src="<%= BASE_URL %>rpay-entry.js"></script>
     <script type="text/javascript" src="https://secure.scan6show.com/js/207145.js"></script>
+    <script type="application/ld+json">
+      {
+        "@context": "https://schema.org",
+        "@type": "Organization",
+        "address": {
+          "@type": "PostalAddress",
+          "addressLocality": "London, UK",
+          "postalCode": "EC1V 2NX",
+          "streetAddress": "Kemp House 152 - 160 City Road"
+        },
+        "email": "info@risidio.com",
+        "name": "Risidio Ltd.",
+        "logo": "https://risidiocorp.cdn.prismic.io/risidiocorp/fa2078f1-300f-4c21-81d8-2d26ff9bb004_risidio_logo_white.svg",
+        "sameAs": [
+          "https://twitter.com/Risidio1",
+          "https://www.instagram.com/risidio_official/",
+          "https://discord.com/invite/sQaKVft"
+        ]
+      }
+    </script>
     <noscript><img src="https://secure.scan6show.com/207145.png" alt="" style="display:none;" /></noscript>
   </head>
   <body>
