+++ src/assets/scss/custom.scss
@@ -156,33 +156,50 @@ body {
   color: #999;
   font-weight: 200;
 }
-.footer{
+.footer {
   /*position: fixed;
   bottom: 0;*/
   width: 100%;
   height: 150px;
   padding: 0;
 }
-h1 {
+footer h3 {
+  text-align: left;
+  font-size: 25px;
+  font-weight: 700;
+  letter-spacing: 0px;
+}
+
+.visually-hidden {
+  position: absolute !important;
+  border: 0 !important;
+  height: 1px !important;
+  width: 1px !important;
+  padding: 0 !important;
+  overflow: hidden !important;
+  clip: rect(0, 0, 0, 0) !important;
+}
+
+h2 {
   text-align: center;
   font-size: 25px;
   font-weight: 700;
   letter-spacing: 0px;
 }
-h2 {
+h3 {
   font-size: 36px;
   font-weight: 200;
   letter-spacing: 0px;
   text-transform: capitalize;
   white-space: normal;
 }
-h3 {
+h4 {
   text-align: left;
   font-size: 21px;
   font-weight: 700;
   letter-spacing: 0px;
 }
-h4 {
+h5 {
   text-align: center;
   font-size: 18px;
   font-weight: 700;
@@ -324,7 +341,7 @@ a:hover {
 }
 
 // SLIDERS
-#section2 .vueperslides__parallax-wrapper, .vueperslides__parallax-wrapper .vueperslides__track 
+#homesection2 .vueperslides__parallax-wrapper, .vueperslides__parallax-wrapper .vueperslides__track 
 { overflow: visible; }
 
 .vueperslides__inner .vueperslides__arrows--outside .vueperslides__arrow--prev { left: -8.5em; }
