+++ src/components/csr/CSRSection2.vue
@@ -1,269 +1,133 @@
 <template>
-  <section class="bg-white">
-    <div class="header-title">
-      <div>
-        <h1 >Our Values</h1>
-      </div>
-      <div>
-        <h2 >For A Better Future</h2>
-      </div>
-    </div>
-            <div class="vueSlideContainer galleryContainer">
-              <vueper-slides
-              :infinite="false"
-              fixed-height="true"
-              class="no-shadow"
-              :arrows="showArrow"
-              :arrows-outside="showArrow"
-              :touchable="touchableSlide"
-              :gap="10"
-              :visible-slides="1"
-              >
-                <template v-if="showArrow == true" #arrow-left>
-                  <img src="../../assets/img/arrow.png" alt="placeholder" class="arrow1"/>
-                </template>
-                <template v-if="showArrow == true" #arrow-right>
-                  <img src="../../assets/img/arrow.png" alt="placeholder" class="arrow2"/>
-                </template>
-                <vueper-slide v-for="(slide, index) of content.section2" :key="index">
-                    <template #content>
-                        <div class="slideContainer">
-                            <div class="slideImage">
-                              <div class="collectionImageBack">
-                                <img :src="slide.image.url" alt="slide" class="slide"/>
-                              </div>
-                              <prismic-rich-text class="slide-text-p" :field="slide.content"/>
-                              <!-- <p class="slide-text-p">{{slide.content[0].text}}</p> -->
-                            </div>
-                        </div>
-                       </template>
-                </vueper-slide>
-              </vueper-slides>
+    <section id="csr-section2" class="bg-lighter">
+        <div class="header-title">
+          <div>
+          <h2> We Care</h2>
+          </div>
+            <div> <h3>Sustainable Development Goals</h3>
             </div>
-      </section>
+            </div>
+            <div>
+              <p class="title">Roadmap and Targets</p>
+              <p class="text">EARTH DAY – APRIL 22, 2022</p>
+              <p class="text" style="max-width: 500px; margin: 0 auto;">Earth Day 2022 marks the beginning of the rollout of our IndigeNFT Genesis collection. The theme for Earth Day 2022 (‘Invest in our Planet’) calls on business to build a more environmentally and socially conscious future for the Earth, a message strongly aligned with our own values.</p>
+            </div>
+            <div class="sustainability">
+    <div>
+        <div>
+            <div class="csrinfo">
+             <img src="../../assets/img/service.png" alt="" class="image"/>
+                <p class="title1"> Governance</p>
+                <p class="text">Corporate Social Responsibility is at the heart of Risidio. Our governing ethos involves expanding our existing ecosystem of partners to establish an environmentally and socially conscious network.</p>
+            </div>
+          </div>
+      </div>
+    <div>
+          <div>
+            <div class="csrinfo">
+             <img src="../../assets/img/services1.png" alt="" class="image"/>
+                <p class="title1"> Strategy</p>
+                <p class="text">To minimise our carbon footprint, we use Stacks’ Proof-of-Transfer (PoX) blockchain mining mechanism for Bitcoin in our NFT marketplace and support the  <a href="https://www.sustainablebtc.org/" style="color: #E9493D;">Sustainable Bitcoin Standard</a>.</p>
+            </div>
+          </div>
+      </div>
+      <div>
+          <div>
+            <div class="csrinfo">
+             <img src="../../assets/img/services2.png" alt="" class="image"/>
+                <p class="title1"> Risk Management</p>
+                <p class="text">We've conducted and published research into a range of CSR projects <a href="https://medium.com/risidio/fake-art-nfts-and-indigenous-rights-be2d8f32deb8"></a> to highlight the fake indigenous art issue and how NFT technologies can support indigenous communities.</p>
+            </div>
+          </div>
+      </div>
+      </div>
+    </section>
 </template>
 
 <script>
-import { VueperSlides, VueperSlide } from 'vueperslides'
-import 'vueperslides/dist/vueperslides.css'
 export default {
   name: 'CSRSection2',
   props: ['viewportDimensions', 'content'],
-
-  components: {
-    VueperSlides,
-    VueperSlide
-    // HomeSearchBar
-  },
-  data: () => ({
-    touchableSlide: false,
-    showArrow: true,
-    slide: [
-      {
-        id: '1',
-        text: 'Collection'
-      },
-      {
-        id: '2',
-        text: 'Collection'
-      },
-      {
-        id: '3',
-        text: 'Collection'
-      }
-    ],
-    return: {
-      resultSet: [],
-      loaded: false,
-      rand: 1
-      // touchableSlide: false
+  data () {
+    return {
     }
-  }),
-  created () {
-    window.addEventListener('resize', this.checkScreen)
-    this.checkScreen()
+  },
+  methods: {
+  },
+  computed: {
   }
 }
 </script>
-
-<style lang="scss" scoped>
-.arrow2{
-  transform: rotate(180deg);
+<style scoped>
+/*
+img {
+  width: 100%;
+  height: auto;
 }
-.bg-white{
+*/
+.bg-lighter{
     min-height: 800px;
-    padding: 10rem 1rem;
+    padding: 50px 15px;
     text-align: center;
 }
-.header-title{
-      text-align: center;
-      margin-top: -40px;
-      gap: 20px;
 
-    }
-  .header-title h1{
-      font-weight:600;
-      font-size: 20px;
-      }
-  .header-title h2{
-      font-weight: 200;
-      font-size: 36px;
-      margin-top: 40px;
-      }
-.if{
-  max-width: 1300px;
+.sustainability {
+  max-width: 1500px;
   margin: auto;
-}
-.slide-text-p{
-    font: normal normal 300 14px/18px Montserrat;
-    max-width: 205px;
-    margin-left: 50px;
-}
-.slide-text-p::v-deep a {
-  color: #E9493D;
-}
-.slideText a{
-  margin-left: 190px;
-  font-size: 12px;
-}
-.text{
-  color: #F9B807;
-}
-.title{
-   margin-left: 300px;
-   margin-top: -200px;
-}
- .title h2{
-   font-weight: 600;
-   font-size: 20px;
- }
-.vueperslide{
-  background-color:rgba(255, 255, 255, 0.637);
-  border-radius: 30px;
-}
-.vueperslides--fixed-height {
-  height: 287px;
-  max-width: 1135px;
-  margin-left: auto;
-  margin-right: auto;
-  margin-bottom: 25px;
-  backdrop-filter: blur(2rem);
-}
-.vueperslides__arrow .arrow1{
-  width: 50px;
-  height: 50px;
-  color: white;
-  position: relative;
-}
-.vueperslides__arrow .arrow2{
-  width: 50px;
-  height: 50px;
-  color: white;
-  position: relative;
-}
-.vueperslides::v-deep .vueperslides__bullets {
-  bottom: -35%;
-}
-.slideContainer{
   display: flex;
   flex-wrap: wrap;
-  justify-content: center;
-  height:100%;
-  padding: 25px 20px;
-  margin-top: 20px;
+  gap: 20px;
+  justify-content: space-evenly;
+  flex-direction: row;
 }
-.collectionImage{
-  display: block;
-  width: 211px;
-  height: 189px;
-  border-radius: 10px;
-  margin: auto;
-  box-shadow: 10px 10px 30px white;
-  object-fit: cover;
-}
-.slideImage{
-  display: flex;
-  align-items: center;
-}
-.collectionImageBack{
+.csrinfo {
   position: relative;
-  display: flex;
+  flex: 0 1 300px;
+  padding: 35px 20px;
+  text-align: center;
+  margin-top: 90px;
 }
-.collectionImageFront{
-  position: absolute;
+.csrinfo p {
+  max-width: 350px;
   margin-left: auto;
   margin-right: auto;
-  left: 0;
-  right: 0;
+}
+.image {
+  position: absolute;
+  top: 10px;
+  right: 0; left: 0;
+  margin: auto;
+  width: 72px;
+  height: 72px;
+}
+.title {
+  font-size: 35px;
+  font-weight: 300;
   text-align: center;
-  transform: scale(1.15)
+  margin-top: 30px;
 }
-@media only screen and (max-width:1330px) {
-  .arrow1{
-    margin-left: 100px;
-  }
-  .arrow2{
-    margin-right: 100px;
-  }
+.title1 {
+  font-size: 35px;
+  font-weight: 300;
+  text-align: center;
+  margin-top: 90px;
 }
-@media only screen and (max-width:1290px) {
-    .vueperslides--fixed-height {
-    max-width: 85%;
-  }
+.text {
+  font-size: 14px;
+  line-height: 18px;
+  text-align: center;
 }
-.notMobileHeader, .mobileHeader{
-    margin-bottom: 1rem;
-    letter-spacing: 1px;
-    font: normal normal 300 30px/55px Montserrat;
-    width: 50%;
+.header-title {
+  text-align: center;
+  gap: 20px;
 }
-.mobileHeader {
-  display: none;
+.header-title h2 {
+  font-weight:600;
+  font-size: 20px;
 }
-.notMobileHeader{
-  display: block;
+.header-title h3 {
+  font-weight: 200;
+  font-size: 36px;
+  margin-top: 40px;
 }
-@media only screen and (max-width: 767px){
-  .collectionImageBack img{
-    width: 200px;
-    height: auto;
-  }
-  .vueperslides--fixed-height {
-    height: 510px;
-  }
-  .vueperslides::v-deep .slideImage {
-    display: flex;
-    flex-direction: column;
-  }
-  .vueperslides::v-deep .collectionImageBack {
-    display: block;
-  }
-  .vueperslides::v-deep .title, .vueperslides::v-deep .innovate, .vueperslides::v-deep .slideText {
-    margin: 0;
-  }
-  .vueperslides::v-deep .slide-text-p, .vueperslides::v-deep .text {
-    margin: 25px 0 0;
-    margin-left: 0;
-    max-width: 200px;
-  }
-  .vueperslides::v-deep .title {
-    margin-top: 10px;
-  }
-  .vueperslides::v-deep .vueperslides__bullets {
-    bottom: 13%;
-  }
-}
-@media only screen and (max-width: 500px) {
-  .arrow1{
-    display: none;
-  }
-  .arrow2{
-    display: none;
-  }
-}
-// @media only screen and (max-width: 389px) {
-//   .vueperslides::v-deep p.slide-text-p, .vueperslides::v-deep .text, .vueperslides::v-deep .slideImage, .vueperslides::v-deep .slideImage img {
-//     width: 100%;
-//   }
-// }
 </style>
