+++ src/components/ourwork/OurWorkSection0.vue
@@ -1,14 +1,14 @@
 <template>
-<section id="section0" :style="bannerImage" class="d-flex justify-content-center align-items-center flex-column text-white">
+<section id="ourwork-section0" :style="bannerImage" class="d-flex justify-content-center align-items-center flex-column text-white">
   <div class="container my-auto pb-5 text-center">
-    <h1 class="mt-5">{{ content.main_title[0].text }}</h1>
+    <h2 class="mt-5">{{ content.main_title[0].text }}</h2>
     <!-- <p class="para">{{ content.subtitle_0[0].text }}</p>
     <a class="text">{{ content.subtitle_1[0].text }}</a> -->
     <!--<h2 class="">{{content.title1[0].text}}</h2>
     <prismic-rich-text class="main-content my-4" :field="content.description"/>
     <div><a href="https://ruma.risidio.com"><button class="">Try now!</button></a></div>-->
   </div>
-  <div v-scroll-to="{ element: '#our-work-section1', duration: 750 }" class="icon-down">
+  <div v-scroll-to="{ element: '#ourwork-section1', duration: 750 }" class="icon-down">
     <i class="fas fa-chevron-down"></i>
   </div>
 </section>
@@ -46,11 +46,8 @@ export default {
 </script>
 
 <style scoped>
-h1 {
-  margin-bottom: 1rem;
-}
 h2 {
-  margin-bottom: 2rem;
+  margin-bottom: 1rem;
 }
 .main-content >>> p {
   text-align: center;
