+++ src/views/AboutUs.vue
@@ -1,8 +1,8 @@
 <template>
 <div v-if="content">
     <about-section0 :viewportDimensions="sectionDimensions" :content="content"/>
-    <about-section2 :viewportDimensions="sectionDimensions" :content="content"/>
     <about-section1 :viewportDimensions="sectionDimensions" :content="content"/>
+    <about-section2 :viewportDimensions="sectionDimensions" :content="content"/>
     <about-section3 :viewportDimensions="sectionDimensions" :content="content"/>
     <section6 :viewportDimensions="sectionDimensions"/>
 </div>
@@ -11,8 +11,8 @@
 <script>
 // @ is an alias to /src
 import AboutSection0 from '@/components/aboutus/AboutSection0'
-import AboutSection2 from '@/components/aboutus/AboutSection2'
 import AboutSection1 from '@/components/aboutus/AboutSection1'
+import AboutSection2 from '@/components/aboutus/AboutSection2'
 import AboutSection3 from '@/components/aboutus/AboutSection3'
 import Section6 from '@/components/home/Section6'
 import { SITE_CONSTANTS } from '@/site-constants'
@@ -22,8 +22,8 @@ export default {
   // mixins: [requestMixins],
   components: {
     AboutSection0,
-    AboutSection2,
     AboutSection1,
+    AboutSection2,
     AboutSection3,
 
     Section6
